Two updates Read carfully so that you know what I added
Added pressureplate to togglerings which responds to pressure.

Added code in tag_gamemode blueprint start event to set the minimum value for culling opacity in the parameter collection "player data". This happens on level load if the level is not equal to the menu, this is determined by its case sensitive level name so do not change the level name for the main menu level from "Menu". This addition means no transparency-dither effect in the editor but intead it is set up automatically ingame when loading each level exept in the main menu.